fix(solve): stop channel stores discarding aliases and injecting spurious chars
InputStoreImpl/OutputStoreImpl.setCurrent(alias) rebuilt the whole store
as just {"$current" -> channel}, discarding every other registered
alias (including the one just switched to). InputChannelFromString
(the JS/Node backend for InputChannel.of(string)) synthesized a
trailing '\n' after every line, even an empty one, so string-backed
input never behaved like the JVM's StringReader-backed channel.

Both were caught by io-lib's I/O predicate tests; those tests are
restored to their cross-platform form now that the underlying bugs
are fixed (one JS-only case is left pinned, tied to an unrelated
NumberTypeTester bug misclassifying negative integers as reals).
